Add portable way of retrieving changetime for resultserver
I thought it was a good idea if developers working on macOS could at
least run "make" locally in order to test their local changes quickly.

Turns out we have a Linux-only code in resultserver, because Go's stat
structure is platform-dependent. This patch therefore splits the
platform dependent code into directory.go that just contains the
structure definition and a helper function and then per-platform
NewDirectory() function that has its own implementations for Linux and
macOS.
