+++ pkg/utils/directory_darwin.go
@@ -1,0 +1,14 @@
+package utils
+
+import (
+	"os"
+	"syscall"
+)
+
+func NewDirectory(path string, info os.FileInfo) Directory {
+	statinfo := info.Sys().(*syscall.Stat_t)
+	return Directory{
+		CreationTime: timespecToTime(statinfo.Ctimespec),
+		Path:         path,
+	}
+}
